Spring Security upd.

+++ src/main/java/com/oceanapplication/ocean/models/Token.java
@@ -15,7 +15,7 @@ public class Token {
 
     @Id
     @GeneratedValue
-    public int id;
+    public Long id;
 
     @Enumerated(EnumType.STRING)
     public TokenType tokenType = TokenType.BEARER;
